+++ R/UTILS-expressions.R
@@ -4,7 +4,17 @@ prepare_expr <- quote({
     if (pairwise && length(x) != length(y))
         stop("Pairwise distances require the same amount of series in 'x' and 'y'.")
     # Get appropriate matrix (big.matrix was needed before)
-    D <- allocate_distmat(length(x), length(y), pairwise, symmetric) # UTILS-utils.R
-    fill_type <- if (pairwise) "PAIRWISE" else if (symmetric) "SYMMETRIC" else "PRIMARY"
+    D <- allocate_distmat(length(x), length(y), pairwise, symmetric, lower_triangular_only, diagonal) # UTILS-utils.R
+
+    fill_type <- if (pairwise) {
+        "PAIRWISE"
+    } else if (lower_triangular_only) {
+        if (diagonal) "LOWER_TRIANGULAR_DIAGONAL" else "LOWER_TRIANGULAR"
+    } else if (symmetric) {
+        "SYMMETRIC"
+    } else {
+        "PRIMARY"
+    }
+
     mat_type <- "R_MATRIX"
 })