+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 5.6.0
 * Update Makevars for ARM version of Windows.
 * Sanitize internal usage of `do.call` to avoid huge backtraces.
+* Support lower triangular `distmat` objects for symmetric distances.
 
 ## Version 5.5.12
 * Remove explicit C++ requirements.